Move line elements to lines.py
